Turned single tool ingredient into a list of tools
The idea is that a Jeweling Station recipe can require the use of multiple
tools.

This commit adds an appropriate top-side inventory for the Jeweling Station
and hooks it up to the existing recipes and plugins. We've also cleaned up
the recipe builder to be more friendly when defining recipes.